+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Changelog
 ### Fixes
 
 - [FIR, Anvil Interop] Fix rank-based binding replacements getting dropped for multi-contribution classes in root graphs.
+- [IR, Anvil Interop] Fix rank-based binding replacements getting dropped for multi-contribution classes in graph extensions.
 
 0.9.2
 -----